sign up not  working

+++ app.py
@@ -1,7 +1,8 @@
 from flask import Flask, render_template, redirect, session, flash
 from flask_debugtoolbar import DebugToolbarExtension
-from  models import connect_db, db, User, Tweet
-from forms import UserForm, TweetForm
+from  models import connect_db, db, User, Feedback
+from forms import UserForm, TweetForm, LoginForm
+from sqlalchemy.exc import IntegrityError
 
 app = Flask(__name__)
 app.config["SQLALCHEMY_DATABASE_URI"] = "postgres:///auth_demo"
@@ -22,7 +23,7 @@ def homepage():
 @app.route('/tweets', methods=['GET', 'POST'])
 def show_tweets():
     if "user_id" not in session:
-        flash("please log in first")
+        flash("please log in first", "danger")
         return redirect('/')
     form = TweetForm()
     all_tweets = Tweet.query.all()
@@ -31,18 +32,21 @@ def show_tweets():
         new_tweet = Tweet(text=text, user_id=session['user_id'])
         db.session.add(new_tweet)
         db.session.commit()
-        flash('tweet created')
+        flash('tweet created', "success")
         return redirect('/tweets')
     return render_template("tweets.html", form=form, tweets = all_tweets)
 
-@app.route('/tweets/<int: id>', methods=['POST'])
+@app.route('/tweets/<int:id>', methods=['POST'])
 def delete_tweet(id):
     """delete tweet"""
+    if 'user_id' not in session:
+        flash("Please log in first", "danger")
+        return redirect('/login')
     tweet = Tweet.query.get_or_404(id)
     if tweet.user_id == session['user_id']:
         db.session.delete(tweet)
         db.session.commit()
-        flash('tweet deleted')
+        flash('tweet deleted', "primary")
         return redirect('/tweets')
     flash('you do not have permission to do that')
     return redirect('/tweets')
@@ -56,23 +60,27 @@ def register_user():
         new_user = User.register(username, password)
         
         db.session.add(new_user)
-        db.session.commit()
+        try: 
+            db.session.commit()
+        except IntegrityError:
+            form.username.errors.append('Username taken. Pick another.')
+            return render_template('register.html', form=form)
         session['user_id'] = new_user.id
-        flash('Welcome! Successfully created your account!')
+        flash('Welcome! Successfully created your account!', "success")
         return redirect('/tweets')
     return render_template('register.html', form=form)
 
 
 @app.route('/login', methods=['GET', 'POST'])
 def login_user():
-    form = UserForm()
+    form = LoginForm()
     if form.validate_on_submit():
         username = form.username.data
         password = form.password.data
 
         user = User.authenticate(User, username, password)
         if user:
-            flash(f'welcome back{user.username}!')
+            flash(f'welcome back{user.username}!', "info")
             session['user_id'] = user.id
             return redirect('/tweets')
         else:
